refactor(desktop): move AI panel state into useAi hook (F5.14)
useAiActions now owns the AI subsystem's state (summary/prompt/touch-up), the
openIdRef/aiCache/mirror-ref landmines, updateAiCache, and the reveal/toast
effects — previously scattered across App.tsx. It returns them by their original
names so loadMessage, useBootstrap, useMailActions, useActionPlan, the keydown/
command contexts, and the render all consume them unchanged. Action bodies are a
verbatim move; the hook is now declared before its consumers.

App.tsx 982 -> 876. 39 e2e + tsc + build green.
