+++ desktop/frontend/src/useAiActions.ts
@@ -1,4 +1,4 @@
-import { useCallback, type MutableRefObject, type RefObject } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { backend, summarizeStream, applyPromptStream, applyBulkPromptStream, threadSummaryStream } from "./api";
 import type { MessageDetail, Prompt } from "./apiTypes";
 import type { ComposeInit } from "./Compose";
@@ -11,10 +11,14 @@ export interface AiCacheEntry {
   lastPromptId?: number;
 }
 
-// useAiActions owns the AI action functions: summarize, draft-reply, touch-up,
-// run-prompt (single + bulk), the per-panel dismiss/dismiss-all, regenerate,
-// and thread-summary. AI STATE + the openIdRef/aiCache/mirror refs stay in App
-// (loadMessage consumes them); this hook receives them via deps. Verbatim move.
+// useAi owns the entire AI subsystem: the panel state (summary / prompt /
+// touch-up), the coupling landmines (openIdRef, the per-message aiCache, and the
+// mirror refs the keydown handler/commands read without stale closures), the
+// reveal/toast effects, and the action functions (summarize, draft-reply,
+// touch-up, run-prompt single+bulk, per-panel dismiss, regenerate, thread
+// summary). App wires only the external inputs below and consumes the returned
+// state/setters/refs by their original names, so loadMessage and the render keep
+// working unchanged. The action bodies are a verbatim move.
 export function useAiActions(deps: {
   detail: MessageDetail | null;
   bulkMode: boolean;
@@ -22,37 +26,115 @@ export function useAiActions(deps: {
   aiEnabled: boolean;
   showToast: (m: string) => void;
   setError: (e: string) => void;
-  setSummary: (v: string | null) => void;
-  setSummarizing: (v: boolean) => void;
-  setSummaryForId: (v: string | null) => void;
-  setPromptResult: (v: string | null) => void;
-  setPromptLabel: (v: string) => void;
-  setPromptRunning: (v: boolean) => void;
-  setPromptForId: (v: string | null) => void;
   setPromptsOpen: (v: boolean) => void;
-  setGeneratingReply: (v: boolean) => void;
-  setTouchUpText: (v: string | null) => void;
-  setTouchingUp: (v: boolean) => void;
   setCompose: (v: ComposeInit | null) => void;
   setBulkPromptLabel: (v: string) => void;
   setBulkPromptText: (v: string | null) => void;
-  openIdRef: MutableRefObject<string | null>;
-  aiCache: MutableRefObject<Map<string, AiCacheEntry>>;
-  updateAiCache: (id: string, patch: Partial<{ summary: string | undefined; touchUp: string | undefined; lastPromptId: number | undefined; promptResults: Record<number, { text: string; label: string }> }>) => void;
-  summaryRef: MutableRefObject<string | null>;
-  promptResultRef: MutableRefObject<string | null>;
-  touchUpTextRef: MutableRefObject<string | null>;
-  promptLabelRef: MutableRefObject<string>;
-  runningLabelRef: MutableRefObject<Record<string, string>>;
-  promptPanelRef: RefObject<HTMLDivElement>;
 }) {
   const {
     detail, bulkMode, selected, aiEnabled, showToast, setError,
-    setSummary, setSummarizing, setSummaryForId, setPromptResult, setPromptLabel, setPromptRunning,
-    setPromptForId, setPromptsOpen, setGeneratingReply, setTouchUpText, setTouchingUp, setCompose,
-    setBulkPromptLabel, setBulkPromptText, openIdRef, aiCache, updateAiCache, summaryRef,
-    promptResultRef, touchUpTextRef, promptLabelRef, runningLabelRef, promptPanelRef,
+    setPromptsOpen, setCompose, setBulkPromptLabel, setBulkPromptText,
   } = deps;
+
+  // Refs to the AI result panels so we can scroll them into view when they
+  // appear — otherwise, if the reader is scrolled down, the panel renders above
+  // the fold and it looks like nothing happened.
+  const summaryPanelRef = useRef<HTMLDivElement>(null);
+  const promptPanelRef = useRef<HTMLDivElement>(null);
+  const touchUpRef = useRef<HTMLDivElement>(null);
+  const [summary, setSummary] = useState<string | null>(null);
+  const [summarizing, setSummarizing] = useState(false);
+  // The message a summary run/result belongs to, so a summary started on one
+  // email doesn't paint its "Generating…" / stream over another you navigated to.
+  const [summaryForId, setSummaryForId] = useState<string | null>(null);
+  const [promptResult, setPromptResult] = useState<string | null>(null);
+  const [promptLabel, setPromptLabel] = useState("");
+  const [promptRunning, setPromptRunning] = useState(false);
+  // The message a single-message prompt result/run belongs to, so a run started
+  // on one email doesn't paint its "Generating…" over a different email you've
+  // since navigated to.
+  const [promptForId, setPromptForId] = useState<string | null>(null);
+  // Always the id of the message currently open in the reader, so a streaming
+  // prompt can tell it should stop updating the visible panel once you move away.
+  const openIdRef = useRef<string | null>(null);
+  const [generatingReply, setGeneratingReply] = useState(false);
+  const [touchUpText, setTouchUpText] = useState<string | null>(null);
+  const [touchingUp, setTouchingUp] = useState(false);
+  // Remember AI results per message (session) so navigating away and back shows
+  // the summary / prompt output / reformat again instead of a blank panel. The
+  // backend also caches, but the frontend state was cleared on every open.
+  const aiCache = useRef<Map<string, AiCacheEntry>>(new Map());
+  // updateAiCache merges a patch into a message's cache entry (creating it if
+  // needed); a key set to undefined deletes it. Consolidates the repeated
+  // get-or-{}/mutate/set dance around aiCache.current.
+  const updateAiCache = useCallback(
+    (
+      id: string,
+      patch: Partial<{
+        summary: string | undefined;
+        touchUp: string | undefined;
+        lastPromptId: number | undefined;
+        promptResults: Record<number, { text: string; label: string }>;
+      }>,
+    ) => {
+      const e = aiCache.current.get(id) ?? {};
+      for (const [k, v] of Object.entries(patch)) {
+        if (v === undefined) delete (e as Record<string, unknown>)[k];
+        else (e as Record<string, unknown>)[k] = v;
+      }
+      aiCache.current.set(id, e);
+    },
+    [],
+  );
+  // Refs mirroring the AI-panel state so the keydown handler / commands can read
+  // fresh values without stale closures (for :dismiss, :regenerate, layered Esc).
+  const summaryRef = useRef(summary);
+  summaryRef.current = summary;
+  const promptResultRef = useRef(promptResult);
+  promptResultRef.current = promptResult;
+  const promptLabelRef = useRef(promptLabel);
+  promptLabelRef.current = promptLabel;
+  const promptRunningRef = useRef(promptRunning);
+  promptRunningRef.current = promptRunning;
+  const promptForIdRef = useRef(promptForId);
+  promptForIdRef.current = promptForId;
+  const summarizingRef = useRef(summarizing);
+  summarizingRef.current = summarizing;
+  const summaryForIdRef = useRef(summaryForId);
+  summaryForIdRef.current = summaryForId;
+  // Label of the prompt currently streaming, keyed by message id, so returning
+  // to a message mid-run can restore its panel title (the global promptLabel is
+  // reset when you navigate away).
+  const runningLabelRef = useRef<Record<string, string>>({});
+  const touchUpTextRef = useRef(touchUpText);
+  touchUpTextRef.current = touchUpText;
+
+  // When an AI result panel starts, reveal it (the panels render at the top of
+  // the reader, so if you'd scrolled down they'd appear above the fold and look
+  // like a no-op) and flash a toast so there's immediate feedback either way.
+  useEffect(() => {
+    if (summarizing && summaryForId && summaryForId === detail?.id) {
+      summaryPanelRef.current?.scrollIntoView({ block: "start", behavior: "smooth" });
+      showToast("Summarizing…");
+    }
+  }, [summarizing, summaryForId, detail?.id, showToast]);
+  useEffect(() => {
+    // Only for a single-message prompt on the message that's actually open (a
+    // bulk run streams into its own modal; promptForId is null for it).
+    if (promptRunning && promptForId && promptForId === detail?.id) {
+      promptPanelRef.current?.scrollIntoView({ block: "start", behavior: "smooth" });
+      showToast(promptLabel ? `Applying ${promptLabel}…` : "Applying prompt…");
+    }
+  }, [promptRunning, promptForId, detail?.id, promptLabel, showToast]);
+  useEffect(() => {
+    if (touchingUp) showToast("Reformatting…");
+  }, [touchingUp, showToast]);
+  useEffect(() => {
+    // The reformatted panel only mounts once the result is set, so reveal it then.
+    if (touchUpText !== null)
+      touchUpRef.current?.scrollIntoView({ block: "start", behavior: "smooth" });
+  }, [touchUpText]);
+
   const summarize = useCallback(async (id: string, force = false) => {
     setSummaryForId(id);
     setSummarizing(true);
@@ -299,6 +381,16 @@ export function useAiActions(deps: {
   }, [detail]);
 
   return {
+    // panel state
+    summary, setSummary, summarizing, summaryForId,
+    promptResult, setPromptResult, promptLabel, setPromptLabel, promptRunning, setPromptRunning, promptForId,
+    generatingReply, touchUpText, setTouchUpText, touchingUp,
+    // DOM refs
+    summaryPanelRef, promptPanelRef, touchUpRef,
+    // landmine refs (shared with loadMessage / keydown / commands)
+    openIdRef, aiCache, runningLabelRef, promptLabelRef, promptForIdRef,
+    promptRunningRef, summarizingRef, summaryForIdRef,
+    // actions
     summarize, generateReply, touchUp, runPrompt, dismissSummary,
     dismissPrompt, dismissTouchUp, dismissAI, regenerateActive, summarizeThread,
   };
